Compute model characteristics

characterize() attempts to report data driven characteristics of a model fit that describe the model. For example, the number of predictors used, the number of terminal nodes, etc. These are interesting characteristics of the model that often reflect complexity of the model.

Usage

characterize(x, ...)

# Default S3 method
characterize(x, ...)

# S3 method for class 'model_fit'
characterize(x, ...)

# S3 method for class 'workflow'
characterize(x, ...)

# S3 method for class 'glmnet'
characterize(x, penalty = 0.001, ...)

# S3 method for class 'cubist'
characterize(x, committees = NULL, ...)

# S3 method for class 'tidy_cubist'
characterize(x, committees = NULL, ...)

# S3 method for class 'cubist'
characterize(x, committees = NULL, ...)

# S3 method for class 'tidy_C50'
characterize(x, trials = max(x$trials), ...)

# S3 method for class 'C5.0'
characterize(x, trials = x$trials["Actual"], ...)

# S3 method for class 'xgb.Booster'
characterize(x, nrounds = x$niter, ...)

# S3 method for class 'tidy_xrf'
characterize(x, penalty = 0.001, ...)

# S3 method for class 'xrf'
characterize(x, penalty = 0.001, ...)

# S3 method for class 'lgb.Booster'
characterize(x, trees = NULL, ...)

# S3 method for class 'fda'
characterize(x, ...)

Arguments

x

An object.

...

Not currently used.

penalty

A numeric parameter value for the amount of regularization/penalty of the model.

committees

The number of Cubist committees to use in the characterization.

trials, nrounds, trees

The number of boosting to use in the characterization.

Value

A tibble with columns .metric, .estimator, and .estimate. If there are no statistics to compute, a zero-row tibble is returned. Integer values are converted to double.

Details

For a list of object types supported, see list_characteristics().

See also

Examples

library(rpart)
cart_mod <- rpart(mpg ~ ., data = mtcars)
characterize(cart_mod)
#> # A tibble: 3 × 3
#>   .metric             .estimator .estimate
#>   <chr>               <chr>          <dbl>
#> 1 num_features_active model              2
#> 2 num_features_input  model             10
#> 3 num_term_nodes      model              3

# When there is nothing to report:
ppr_mod <- ppr(mpg ~ ., data = mtcars, nterms = 5)
characterize(ppr_mod)
#> # A tibble: 1 × 3
#>   .metric            .estimator .estimate
#>   <chr>              <chr>          <dbl>
#> 1 num_features_input model             10