Model characterizations across multiple configurations

Some models can make predictions for multiple configurations using the same model object. For example, many boosted tree models using m trees can also make predictions when fewer than m trees were used. These are sometimes called "sub-model parameters".

For such a model, multi_characterize() can produce multiple results from the same object. If the model does not allow for this, the results of characterize() are returned.

Usage

multi_characterize(object, ...)

# Default S3 method
multi_characterize(object, ...)

# S3 method for class 'bundle'
multi_characterize(object, ...)

# S3 method for class 'workflow'
multi_characterize(object, ...)

# S3 method for class 'model_fit'
multi_characterize(object, ...)

# S3 method for class '`_elnet`'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class 'glmnet'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class '`_multnet`'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class 'multnet'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class '`_lognet`'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class 'lognet'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class '`_fishnet`'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class 'fishnet'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class 'cubist'
multi_characterize(object, committees = object$committees, ...)

# S3 method for class 'C5.0'
multi_characterize(object, trials = object$trials["Actual"], ...)

# S3 method for class 'xgb.Booster'
multi_characterize(object, nrounds = object$niter, ...)

# S3 method for class 'xrf'
multi_characterize(object, penalty = NULL, ...)

# S3 method for class 'lgb_trees'
multi_characterize(object, trees = NULL, ...)

# S3 method for class 'lgb.Booster'
multi_characterize(object, trees = NULL, ...)

Arguments

object

A model_fit object.

...

Optional arguments (not currently used)

penalty

A numeric parameter value for the amount of regularization/penalty of the model.

committees

The number of Cubist committees to use in the characterization.

trials, nrounds, trees

The number of boosting to use in the characterization.

Value

A tibble of characterized results (results) with extra columns for the sub-model parameter. If there is no specific multi_characterize() method, there are not parameter columns and the results column has the results of using a basic characterize() call.

Details

For glmnet models, there are methods for the different classes of that model (a.k.a. lognet class for Poisson models) as well as a variation of that class (_lognet). The latter case corresponds to the object produced by the parsnip package for that model.

See also

Examples

if (FALSE) { # characterize:::is_rcmd_check()
if (rlang::is_installed(c("modeldata", "ggplot2", "Cubist"))) {
  library(Cubist)
  library(dplyr)
  library(ggplot2)
  library(tidyr)

  # ----------------------------------------------------------------------------

  data(ames, package = "modeldata")
  ames$Sale_Price <- log10(ames$Sale_Price)

  cubist_fit <-
    cubist(x = ames %>% select(-Sale_Price), y = ames$Sale_Price, committees = 5)

  # ----------------------------------------------------------------------------
  # Get results for a specific model configuration

  characterize(cubist_fit, committees = 3)

  # ----------------------------------------------------------------------------
  # Get results for many model configurations

  res <- multi_characterize(cubist_fit, committees = 1:5)
  res

  # unnest to use them:
  res %>%
    unnest(cols = results) %>%
    ggplot(aes(committees, .estimate)) +
    geom_point() +
    facet_wrap(~ .metric, scale = "free_y")
}
}