tune helpers

tune helpers

Usage

retain_characteristics(wflow)

collect_characteristics(
  x,
  summarize = TRUE,
  parameters = NULL,
  add_metrics = FALSE,
  wide = FALSE
)

Arguments

wflow

A fitted model workflow.

x

The results of tune_grid(), tune_bayes(), fit_resamples(), or last_fit(). For collect_predictions(), the control option save_pred = TRUE should have been used.

summarize

A logical; should metrics be summarized over resamples (TRUE) or return the values for each individual resample. Note that, if x is created by last_fit(), summarize has no effect. For the other object types, the method of summarizing predictions is detailed below.

parameters

An optional tibble of tuning parameter values that can be used to filter the predicted values before processing. This tibble should only have columns for each tuning parameter identifier (e.g. "my_param" if tune("my_param") was used).

add_metrics

A logical to combine the results the associated performance metrics.

wide

A logical for making some information in wide format.

Value

A tibble.

Details

When using the tune or finetune packages, retain_characteristics() can be used with the resampling and tuning functions to return the characteristics from multiple model configurations. These results are contained in a column called .extracts.

collect_characteristics() can be used to easily wrangle the results in the .extracts column into a tibble of usable results.

The vignette Using characterize with tidymodels shows an example.